requirements.txt
# Core Django and REST framework
Django==5.2.5
djangorestframework==3.16.1
asgiref==3.9.1
sqlparse==0.5.3

# Database & timezone support
pytz==2025.2
tzdata==2025.2

# Data processing & ML
numpy==2.3.2
pandas==2.3.2
joblib==1.5.1
scikit-learn==1.7.1
python-dateutil==2.9.0.post0

# NLP embeddings
sentence-transformers==5.1.1

# Standard utility libraries
six==1.17.0
threadpoolctl==3.6.0
packaging==25.0